BaseConjunta/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Brotli==1.1.0
dj-database-url==2.1.0
Django==4.1.7
django-cors-headers==3.14.0
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==21.2.0
importlib-metadata==6.1.0
Markdown==3.4.3
packaging==23.2
Pillow==10.0.1
psycopg2-binary==2.9.9
PyJWT==2.6.0
pytz==2023.3
sqlparse==0.4.3
typing_extensions==4.8.0
tzdata==2023.3
whitenoise==6.6.0
zipp==3.15.0